# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  nombre: string;
  email: string;
  /** @minLength 8 */
  password: string;
  telefono?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type PublicUserRol = (typeof PublicUserRol)[keyof typeof PublicUserRol];

export const PublicUserRol = {
  cliente: "cliente",
  admin: "admin",
} as const;

export interface PublicUser {
  id: string;
  nombre: string;
  email: string;
  rol: PublicUserRol;
  telefono?: string | null;
  activo: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: PublicUser;
}

export type RoomTipo = (typeof RoomTipo)[keyof typeof RoomTipo];

export const RoomTipo = {
  sencilla: "sencilla",
  doble: "doble",
  suite: "suite",
  cabana: "cabana",
} as const;

export interface Room {
  id: string;
  nombre: string;
  descripcion?: string | null;
  tipo: RoomTipo;
  precioNoche: string;
  capacidad: number;
  imageUrl?: string | null;
  amenidades?: string | null;
  activo: boolean;
  createdAt: string;
}

export type RoomInputTipo = (typeof RoomInputTipo)[keyof typeof RoomInputTipo];

export const RoomInputTipo = {
  sencilla: "sencilla",
  doble: "doble",
  suite: "suite",
  cabana: "cabana",
} as const;

export interface RoomInput {
  nombre: string;
  descripcion?: string;
  tipo: RoomInputTipo;
  precioNoche: string;
  capacidad: number;
  imageUrl?: string;
  amenidades?: string;
  activo?: boolean;
}

export interface ReservationInput {
  habitacionId: string;
  fechaEntrada: string;
  fechaSalida: string;
  notas?: string;
}

export type ReservationWithDetailsEstado =
  (typeof ReservationWithDetailsEstado)[keyof typeof ReservationWithDetailsEstado];

export const ReservationWithDetailsEstado = {
  confirmada: "confirmada",
  cancelada: "cancelada",
  completada: "completada",
} as const;

export interface ReservationWithDetails {
  id: string;
  clienteId: string;
  habitacionId: string;
  fechaEntrada: string;
  fechaSalida: string;
  precioTotal: string;
  estado: ReservationWithDetailsEstado;
  notas?: string | null;
  createdAt: string;
  room: Room;
  cliente: PublicUser;
}

export interface AdminStats {
  totalReservaciones: number;
  reservacionesHoy: number;
  habitacionesDisponibles: number;
  totalHabitaciones: number;
  ingresosMes: number;
  ocupacionPorcentaje: number;
  reservasRecientes: ReservationWithDetails[];
}

export type ListRoomsParams = {
  fechaEntrada?: string;
  fechaSalida?: string;
  capacidad?: number;
  tipo?: string;
};

export type ListReservationsParams = {
  estado?: string;
};