# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sistema de Reservas Hotel - API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: rooms
    description: Room management
  - name: reservations
    description: Reservation management
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List available rooms
      parameters:
        - name: fechaEntrada
          in: query
          schema:
            type: string
        - name: fechaSalida
          in: query
          schema:
            type: string
        - name: capacidad
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      security:
        - bearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservationWithDetails"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationWithDetails"
        "400":
          description: Room not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get reservation by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationWithDetails"
    patch:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel a reservation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationWithDetails"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    RegisterInput:
      type: object
      properties:
        nombre:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        telefono:
          type: string
      required:
        - nombre
        - email
        - password

    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    PublicUser:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        email:
          type: string
        rol:
          type: string
          enum: [cliente, admin]
        telefono:
          type: string
          nullable: true
        activo:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - nombre
        - email
        - rol
        - activo
        - createdAt

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/PublicUser"
      required:
        - token
        - user

    Room:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
          nullable: true
        tipo:
          type: string
          enum: [sencilla, doble, suite, cabana]
        precioNoche:
          type: string
        capacidad:
          type: integer
        imageUrl:
          type: string
          nullable: true
        amenidades:
          type: string
          nullable: true
        activo:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - nombre
        - tipo
        - precioNoche
        - capacidad
        - activo
        - createdAt

    RoomInput:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        tipo:
          type: string
          enum: [sencilla, doble, suite, cabana]
        precioNoche:
          type: string
        capacidad:
          type: integer
        imageUrl:
          type: string
        amenidades:
          type: string
        activo:
          type: boolean
      required:
        - nombre
        - tipo
        - precioNoche
        - capacidad

    ReservationInput:
      type: object
      properties:
        habitacionId:
          type: string
        fechaEntrada:
          type: string
          format: date
        fechaSalida:
          type: string
          format: date
        notas:
          type: string
      required:
        - habitacionId
        - fechaEntrada
        - fechaSalida

    ReservationWithDetails:
      type: object
      properties:
        id:
          type: string
        clienteId:
          type: string
        habitacionId:
          type: string
        fechaEntrada:
          type: string
        fechaSalida:
          type: string
        precioTotal:
          type: string
        estado:
          type: string
          enum: [confirmada, cancelada, completada]
        notas:
          type: string
          nullable: true
        createdAt:
          type: string
        room:
          $ref: "#/components/schemas/Room"
        cliente:
          $ref: "#/components/schemas/PublicUser"
      required:
        - id
        - clienteId
        - habitacionId
        - fechaEntrada
        - fechaSalida
        - precioTotal
        - estado
        - createdAt
        - room
        - cliente

    AdminStats:
      type: object
      properties:
        totalReservaciones:
          type: integer
        reservacionesHoy:
          type: integer
        habitacionesDisponibles:
          type: integer
        totalHabitaciones:
          type: integer
        ingresosMes:
          type: number
        ocupacionPorcentaje:
          type: number
        reservasRecientes:
          type: array
          items:
            $ref: "#/components/schemas/ReservationWithDetails"
      required:
        - totalReservaciones
        - reservacionesHoy
        - habitacionesDisponibles
        - totalHabitaciones
        - ingresosMes
        - ocupacionPorcentaje
        - reservasRecientes