# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationWithDetails } from "./reservationWithDetails";

export interface AdminStats {
  totalReservaciones: number;
  reservacionesHoy: number;
  habitacionesDisponibles: number;
  totalHabitaciones: number;
  ingresosMes: number;
  ocupacionPorcentaje: number;
  reservasRecientes: ReservationWithDetails[];
}