# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUserRol } from "./publicUserRol";

export interface PublicUser {
  id: string;
  nombre: string;
  email: string;
  rol: PublicUserRol;
  telefono?: string | null;
  activo: boolean;
  createdAt: string;
}