# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  nombre: string;
  email: string;
  /** @minLength 8 */
  password: string;
  telefono?: string;
}