# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  habitacionId: string;
  fechaEntrada: Date;
  fechaSalida: Date;
  notas?: string;
}