# lib/api-zod/src/generated/types/reservationWithDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUser } from "./publicUser";
import type { ReservationWithDetailsEstado } from "./reservationWithDetailsEstado";
import type { Room } from "./room";

export interface ReservationWithDetails {
  id: string;
  clienteId: string;
  habitacionId: string;
  fechaEntrada: string;
  fechaSalida: string;
  precioTotal: string;
  estado: ReservationWithDetailsEstado;
  notas?: string | null;
  createdAt: string;
  room: Room;
  cliente: PublicUser;
}