# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sistema de Reservas Hotel - API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomTipo } from "./roomTipo";

export interface Room {
  id: string;
  nombre: string;
  descripcion?: string | null;
  tipo: RoomTipo;
  precioNoche: string;
  capacidad: number;
  imageUrl?: string | null;
  amenidades?: string | null;
  activo: boolean;
  createdAt: string;
}